requirements.txt
streamlit>=1.28.0
tensorflow>=2.12.0,<2.15.0
keras>=2.12.0,<2.15.0
numpy>=1.23.5,<1.25.0
Pillow>=9.5.0,<10.0.0
pandas>=1.5.0,<2.1.0
openpyxl>=3.0.0,<4.0.0
